Push test env. classes on sub-package level
