+++ server/src/test/java/io/spine/server/aggregate/AggregatePartTest.java
@@ -24,11 +24,11 @@ import com.google.common.testing.NullPointerTester;
 import com.google.protobuf.Message;
 import io.spine.core.CommandEnvelope;
 import io.spine.server.BoundedContext;
-import io.spine.server.aggregate.given.AggregatePartTestEnv.AnAggregateRoot;
-import io.spine.server.aggregate.given.AggregatePartTestEnv.TaskDescriptionPart;
-import io.spine.server.aggregate.given.AggregatePartTestEnv.TaskDescriptionRepository;
-import io.spine.server.aggregate.given.AggregatePartTestEnv.TaskPart;
-import io.spine.server.aggregate.given.AggregatePartTestEnv.TaskRepository;
+import io.spine.server.aggregate.given.part.AnAggregateRoot;
+import io.spine.server.aggregate.given.part.TaskDescriptionPart;
+import io.spine.server.aggregate.given.part.TaskDescriptionRepository;
+import io.spine.server.aggregate.given.part.TaskPart;
+import io.spine.server.aggregate.given.part.TaskRepository;
 import io.spine.test.aggregate.ProjectId;
 import io.spine.test.aggregate.Task;
 import io.spine.test.aggregate.command.AggAddTask;
