+++ server/src/test/java/io/spine/server/aggregate/given/part/WrongAggregatePart.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2018, TeamDev. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.server.aggregate.given.part;
+
+import com.google.protobuf.StringValue;
+import io.spine.server.aggregate.AggregatePart;
+import io.spine.validate.StringValueVBuilder;
+
+/**
+ * A class which extends {@link AggregatePart} but does not provide suitable constructor.
+ *
+ * @author Alexander Yevsyukov
+ */
+public class WrongAggregatePart
+        extends AggregatePart<String,
+                              StringValue,
+                              StringValueVBuilder,
+                              AnAggregateRoot> {
+
+    @SuppressWarnings("ConstantConditions")
+    // Supply a "wrong" parameters on purpose to cause the validation failure
+    protected WrongAggregatePart() {
+        super(null);
+    }
+}
